Move 'probably equal' methods to librustc_parse
This is preparation for PR #73084
